Added a check_term callback.
The purpose of this callback is to allow the client to process work
asynchronously. If the client is doing work asynchronously then it can
appear to libcircle that there is no work left to be done and it will
exit before the asynchronous work is done. This function can be used to
let libcircle know that even though the work queue is empty there is
still work being done and that work may add items to the work queue so
don't exit just yet.